requirements.txt
flask
python-dotenv
psycopg2-binary
Flask-SQLAlchemy
Flask-Migrate
werkzeug
uuid
flask-jwt-extended
flask-cors
